+++ apps/server/src/shared/connect/contract.ts
@@ -1,3 +1,5 @@
+import type { PostMessageEventData } from "@repo/utils";
+
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export interface Credential {
   guid: string; //
@@ -65,6 +67,7 @@ export interface Member {
   aggregated_at?: string | null;
   background_aggregation_is_disabled?: boolean;
   connection_status?: string | null | number; //
+  postMessageEventData?: PostMessageEventData;
   guid: string;
   id?: string | null;
   institution_code?: string | null;
@@ -86,6 +89,7 @@ export interface Member {
   skip_aggregation?: boolean | null;
   credentials?: Credential[];
 }
+
 export interface MemberResponse {
   member: Member;
 }