+++ packages/sophtron-adapter/src/adapter.test.ts
@@ -382,6 +382,17 @@ describe("sophtron adapter", () => {
   });
 
   describe("GetConnectionStatus", () => {
+    const createPostMessageEventData = ({ rawStatus, selectedAccountId }) => ({
+      memberConnected: {
+        rawStatus,
+        selectedAccountId,
+      },
+      memberStatusUpdate: {
+        rawStatus,
+        selectedAccountId,
+      },
+    });
+
     it("returns the connection using the memberId and userId if there is no jobId", async () => {
       const response = await adapter.GetConnectionStatus(
         testId,
@@ -418,6 +429,11 @@ describe("sophtron adapter", () => {
 
       expect(response).toEqual({
         id: testUserInstitutionId,
+        challenges: null,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "success",
+          selectedAccountId: "",
+        }),
         userId: testUserId,
         cur_job_id: testJobId,
         status: ConnectionStatus.CONNECTED,
@@ -447,6 +463,11 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        challenges: null,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "failed",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.FAILED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -475,6 +496,11 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        challenges: null,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "AccountsReady",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CREATED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -519,6 +545,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "AccountsReady",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -561,6 +591,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "SecurityQuestion",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -604,6 +638,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "TokenMethod",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -648,6 +686,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "TokenInput",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -689,6 +731,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "TokenInput",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -732,6 +778,10 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "TokenRead",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
@@ -770,36 +820,45 @@ describe("sophtron adapter", () => {
         id: testUserInstitutionId,
         userId: testUserId,
         cur_job_id: testJobId,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: "CaptchaImage",
+          selectedAccountId: "",
+        }),
         status: ConnectionStatus.CHALLENGED,
         aggregator: SOPHTRON_ADAPTER_NAME,
       });
     });
-  });
 
-  it("defaults to connection status created", async () => {
-    server.use(
-      http.get(SOPHTRON_GET_JOB_INFO_PATH, () =>
-        HttpResponse.json({
-          JobID: testJobId,
-          JobType: "agg",
-          UserInstitutionID: testUserInstitutionId,
+    it("defaults to connection status created", async () => {
+      server.use(
+        http.get(SOPHTRON_GET_JOB_INFO_PATH, () =>
+          HttpResponse.json({
+            JobID: testJobId,
+            JobType: "agg",
+            UserInstitutionID: testUserInstitutionId,
+          }),
+        ),
+      );
+
+      const response = await adapter.GetConnectionStatus(
+        testId,
+        testJobId,
+        false,
+        testUserId,
+      );
+
+      expect(response).toEqual({
+        id: testUserInstitutionId,
+        userId: testUserId,
+        cur_job_id: testJobId,
+        challenges: null,
+        postMessageEventData: createPostMessageEventData({
+          rawStatus: undefined,
+          selectedAccountId: "",
         }),
-      ),
-    );
-
-    const response = await adapter.GetConnectionStatus(
-      testId,
-      testJobId,
-      false,
-      testUserId,
-    );
-
-    expect(response).toEqual({
-      id: testUserInstitutionId,
-      userId: testUserId,
-      cur_job_id: testJobId,
-      status: ConnectionStatus.CREATED,
-      aggregator: SOPHTRON_ADAPTER_NAME,
+        status: ConnectionStatus.CREATED,
+        aggregator: SOPHTRON_ADAPTER_NAME,
+      });
     });
   });