+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- added raw_status in connection object which will later made available through client side events, the raw_status field contains the status returned from aggregator before mapped to universal widget contract, this richer information enables client to track the connection status at a more detailed and accurate level which in turns provide the opportunity to trigger more process at a better timing, as well as make diagnosing easier
+- Now passing custom events to the connect widget for member status update and member connected events. Added some sophtron specific values for member status update.
 
 ## [1.0.9]