+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.10]
+
+### Changed
+
+- Now passing custom events to the connect widget for member status update and member connected events. Added some sophtron specific values for member status update.
+
 ## [1.0.9]
 
 ### Fixed